ВСЁ ГОТОВО! ОТЕТ ДОДЕЛАН И ИСПРАВЛЕН! КОД ИСПРАВЛЕН :) EVERYTHING WORKS

+++ sources/SubDirrrProject/libCPP/bankcpp.h
@@ -4,7 +4,7 @@
 class BankCPP
 {
 public:
-    static float compoundInterest(float, float);
+    static double compoundInterest(double, double);
 };
 
 #endif // BANKCPP_H
